Extract graph node types into a submodule
